requirements.txt
numpy
pandas
matplotlib
pillow
requests
emoji
